+++ examples/update/step1/index.ts
@@ -1,12 +1,17 @@
 // Copyright 2016-2021, Pulumi Corporation.
 
+import * as pulumi from "@pulumi/pulumi";
 import * as aws from "@pulumi/aws-native";
 
-const logGroup = new aws.s3.Bucket("bucket", {
+const bucket = new aws.s3.Bucket("bucket", {
   tags: [
     {
       key: "foo",
       value: "bar",
+    },
+    {
+      key: "secretfoo",
+      value: pulumi.secret("secretbar"),
     },
   ],
 });