+++ examples/update/step2/index.ts
@@ -1,12 +1,17 @@
 // Copyright 2016-2021, Pulumi Corporation.
 
+import * as pulumi from "@pulumi/pulumi";
 import * as aws from "@pulumi/aws-native";
 
-const logGroup = new aws.s3.Bucket("bucket", {
+const bucket = new aws.s3.Bucket("bucket", {
   tags: [
     {
       key: "foo",
       value: "buzz", // <-- this value has changed
+    },
+    {
+      key: "secretfoo",
+      value: pulumi.secret("secretbuzz"), // <-- this value has changed
     },
   ],
 });
@@ -41,3 +46,5 @@ const func = new aws.lambda.Function("function", {
   `,
   },
 });
+
+export const secretValue = bucket.tags.apply((tags) => tags!.find((t) => t.key === "secretfoo")?.value);